refactor(connect): extract vLLM preflight into dedicated helper module
