+++ src/lib/actions/sandbox/connect-vllm-preflight.ts
@@ -1,0 +1,24 @@
+// SPDX-FileCopyrightText: Copyright (c) 2026 NVIDIA CORPORATION & AFFILIATES. All rights reserved.
+// SPDX-License-Identifier: Apache-2.0
+
+import { CLI_NAME } from "../../cli/branding";
+import { preflightVllmModelEnv } from "../../inference/vllm-models";
+
+// `NEMOCLAW_VLLM_MODEL` only steers the express-vLLM install path, but users
+// often re-export it in the same shell they later run `connect` in. Run the
+// installer's validators up-front so a typo or a gated model with no
+// `HF_TOKEN` fails fast on the host — before any sandbox readiness probe,
+// inference-route reset, or SSH attach — instead of being silently ignored.
+export function preflightVllmModelEnvOrExit(): void {
+  const result = preflightVllmModelEnv();
+  if (result.ok) return;
+  console.error("");
+  console.error(`  Error: ${result.message}`);
+  console.error(
+    `  Hint: NEMOCLAW_VLLM_MODEL is consumed by the managed-vLLM install path, not \`${CLI_NAME} <name> connect\`.`,
+  );
+  console.error(
+    "  Unset NEMOCLAW_VLLM_MODEL before reconnecting, or fix the value (and token) and re-run the install path that serves the model.",
+  );
+  process.exit(1);
+}
